Handle empty record types.
1. Existing code defines btraverse with an appropriate mixture of fmaps and
nested btraverses; but in the case of an empty record type it needs to call
pure. (The equivalent of the "pure []" case for regular traverse on a list.)
2. A barbie made from an empty record type leaves the clothing parameter "h"
unconstrained. E.g. "data EmptyRecord sw h = EmptyRecord {}". It looks like
GHC is having trouble with inferring the kind of h. In these changes I added a
kind annotation (h :: * -> *). An alternative would be to require PolyKinds in
any client code that tries to define a barbie for an empty record type. I
believe that any non-empty barbies would already be constraining (h :: * -> *)
implicitly.